Refactor land and building property models to use appropriate data types for RightOfWay and update related properties for consistency

+++ Modules/Appraisal/Appraisal/Application/Features/Appraisals/CreateLandProperty/CreateLandPropertyCommand.cs
@@ -45,7 +45,7 @@ public record CreateLandPropertyCommand(
     decimal? SoilLevel = null,
     // Road Access
     decimal? AccessRoadWidth = null,
-    decimal? RightOfWay = null,
+    short? RightOfWay = null,
     decimal? RoadFrontage = null,
     int? NumberOfSidesFacingRoad = null,
     string? RoadPassInFrontOfLand = null,
